Refactor configs, petals.client, petals.bloom
