+++ src/petals/bloom/__init__.py
@@ -1,0 +1,6 @@
+from petals.bloom.modeling_bloom import (
+    DistributedBloomConfig,
+    DistributedBloomForCausalLM,
+    DistributedBloomForSequenceClassification,
+    DistributedBloomModel,
+)
