+++ src/petals/client/modeling_utils.py
@@ -7,7 +7,8 @@ import torch.nn.functional as F
 import torch.utils.checkpoint
 from hivemind import get_logger
 from torch import nn
-from transformers import BloomConfig
+
+from petals.client.from_pretrained import DistributedPretrainedConfig
 
 logger = get_logger(__name__)
 
@@ -19,7 +20,7 @@ class LMHead(nn.Module):
     In addition, it provides an effcient way to deal with half-precision word embeddings on CPU.
     """
 
-    def __init__(self, config: BloomConfig, word_embeddings: nn.Embedding):
+    def __init__(self, config: DistributedPretrainedConfig, word_embeddings: nn.Embedding):
         super().__init__()
         self.word_embeddings = word_embeddings
