+++ src/petals/llama/block.py
@@ -3,7 +3,6 @@ LLaMA intermediate layer
 Based on https://github.com/huggingface/transformers/blob/main/src/transformers/models/llama/modeling_llama.py
 See commit history for authorship.
 """
-import os
 from typing import Optional, Tuple
 
 import torch
